# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
        "title": "kubetabs API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/kubetabs/blob/master/LICENSE.md"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/sysUser": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "用户"
                ],
                "summary": "获取用户角色和职位",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.SysUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/sysUser/{userId}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "用户"
                ],
                "summary": "获取用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户编码",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sysUserList": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "用户"
                ],
                "summary": "列表数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/sysuser/{userId}": {
            "put": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改用户数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.SysUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"修改失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "用户"
                ],
                "summary": "删除用户数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/profile": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "个人中心"
                ],
                "summary": "获取当前登录用户",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/profileAvatar": {
            "post": {
                "description": "获取JSON",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "修改头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "获取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "Add account",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"成功退出系统\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Login": {
            "type": "object",
            "required": [
                "code",
                "password",
                "username",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集",
                    "type": "object"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string"
                }
            }
        },
...
        "models.SysUser": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createBy": {
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "deptId": {
                    "description": "部门编码",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "isDel": {
                    "type": "string"
                },
                "nickName": {
                    "description": "昵称",
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "postId": {
                    "description": "职位编码",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色编码",
                    "type": "integer"
                },
                "salt": {
                    "description": "盐",
                    "type": "string"
                },
                "sex": {
                    "description": "性别",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateTime": {
                    "description": "修改时间",
                    "type": "string"
                },
                "userId": {
                    "description": "编码",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}